refactor: 기존 render 함수 관심사 분리
