+++ src/routes/utils/moveToPage.js
@@ -1,0 +1,9 @@
+import { routerState } from "../routerManager";
+import { routerRender } from "../routerRender";
+
+export const moveToPage = (page) => {
+  if (routerState.mode !== "hash") {
+    window.history.pushState(null, null, `${routerState.basePath}${page}`);
+    routerRender();
+  }
+};
